+++ domain/src/sharedTest/java/br/com/pebmed/domain/FakeUserModel.kt
@@ -1,0 +1,13 @@
+package br.com.pebmed.domain
+
+import br.com.pebmed.domain.entities.UserModel
+
+class FakeUserModel {
+    
+    companion object {
+        fun loadUser() = UserModel(
+            login = "luis.fernandez",
+            avatarUrl = "http://avatar.url"
+        )
+    }
+}
