+++ domain/src/sharedTest/java/br/com/pebmed/domain/usecases/UsefulObjects.kt
@@ -4,6 +4,7 @@ import br.com.pebmed.domain.entities.PullRequestModel
 import br.com.pebmed.domain.entities.RepoModel
 import br.com.pebmed.domain.entities.UserModel
 import br.com.pebmed.domain.extensions.fromJsonGeneric
+import br.com.pebmed.domain.usecases.GetPullRequestsUseCase
 import com.google.gson.Gson
 import java.io.IOException
 import java.nio.file.Files
@@ -17,28 +18,7 @@ object UsefulObjects {
         avatarUrl = "http://avatar.url"
     )
 
-    fun loadPullRequest(user: UserModel) = PullRequestModel(
-        number = 1,
-        htmlUrl = "http://the.url",
-        title = "Title",
-        user = this.loadUser(),
-        body = "Body",
-        createdAt = Date(),
-        comments = 1,
-        commits = 1,
-        additions = 1,
-        deletions = 1,
-        changedFiles = 1
-    )
 
-    fun loadListPullRequestsUseCaseParams() = GetPullRequestsUseCase.Params(
-        owner = "OwnerModel",
-        repoName = "RepoName"
-    )
-
-    fun loadRepos(): List<RepoModel> {
-        return Gson().fromJsonGeneric(readJsonFile("repos.json"))
-    }
 
     @Throws(IOException::class)
     fun readJsonFile(filename: String): String {
