+++ domain/src/test/java/br/com/pebmed/domain/usecases/GetReposUseCaseTest.kt
@@ -1,5 +1,6 @@
 package br.com.pebmed.domain.usecases
 
+import br.com.pebmed.domain.FakeGitRepoModel
 import br.com.pebmed.domain.base.ResultWrapper
 import br.com.pebmed.domain.repository.RepoRepository
 import io.mockk.MockKAnnotations
@@ -33,7 +34,7 @@ class GetReposUseCaseTest {
                 language = fakeLanguage
             )
         } returns ResultWrapper(
-            success = UsefulObjects.loadRepos()
+            success = FakeGitRepoModel.loadRepos()
         )
 
         GetReposUseCase(repoRepository).runAsync(GetReposUseCase.Params(true))
